+++ .changeset/silent-eels-think.md
@@ -1,0 +1,6 @@
+---
+'@ai-sdk/provider-utils': patch
+'ai': patch
+---
+
+chore(provider-utils): move tool-call + tool-result to ai package
