+++ packages/provider-utils/src/index.ts
@@ -24,6 +24,3 @@ export * from './uint8-utils';
 export * from './validate-types';
 export * from './validator';
 export * from './without-trailing-slash';
-
-export type { ToolCall } from './types/tool-call';
-export type { ToolResult } from './types/tool-result';
